relationship between user and listing added

+++ routes/web.php
@@ -20,39 +20,39 @@ Route::get('/', [ListingController::class, 'index']);
 
 //show creat form
 
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //store listing data
 
-Route::post('/listings', [ListingController::class,'store']);
+Route::post('/listings', [ListingController::class,'store'])->middleware('auth');
 
 //show edit form
-Route::get('/listings/{listing}/edit',[ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit',[ListingController::class, 'edit'])->middleware('auth');
 
 //submit the edit form to update
-Route::put('/listings/{listing}',[ListingController::class, 'update']);
+Route::put('/listings/{listing}',[ListingController::class, 'update'])->middleware('auth');
 
 
 //Delete
-Route::delete('/listings/{listing}',[ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}',[ListingController::class, 'destroy'])->middleware('auth');
 
 //single listing
 
 Route::get('/listings/{listing}', [ListingController::class,'show']);
 
-//show register formR
+//show register form
 Route::get('/register', [UserController::class, 'create']
-);
+)->middleware('guest');
 
 //create a new user
 
 Route::post('/users', [UserController::class,'store']);
 //logout
 Route::post
-('/logout', [UserController::class, 'logout']);
+('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 //login form
-Route::get('/login',[UserController::class, 'login']);
+Route::get('/login',[UserController::class, 'login'])->name('login')->middleware('guest');
 //login users
 
 Route::post('/users/authenticate', [UserController::class,'authenticate']);
